pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "DeepSeek Telegram Bot"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
]
